backend/requirements.txt
# Flask and extensions
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
python-dotenv==1.0.0

# Database
SQLAlchemy==2.0.21
psycopg2-binary==2.9.9  # PostgreSQL adapter (for production)

# AWS Bedrock
boto3==1.28.85
strands==0.1.6

# API client
requests==2.31.0

# Utilities
gunicorn==21.2.0  # Production WSGI server
python-json-logger==2.0.7  # Structured logging

# Development
pytest==7.4.2
pytest-cov==4.1.0
